+++ src/models/character/CharacterStats.ts
@@ -10,6 +10,7 @@ class CharacterStats {
      */
     constructor(character: Character) {
         const relicStatProperties: StatPropertyValue[] = character.relics.map(r => [{ statProperty: r.mainStat.mainStatData.statProperty, value: r.mainStat.value }, ...r.subStats.map(s => { return { statProperty: s.subStatData.statProperty, value: s.value }; })]).reduce((a, b) => [...a, ...b]);
+        const lightConeStatProperties: StatPropertyValue[] = [...(character.lightCone?.basicStats ?? []), ...(character.lightCone?.extraStats ?? [])];
         // TODO: add stat properties
     }
 }